// src/platform/packages/shared/cloud/connection_details/types.d.ts
import type { ApiKey } from './tabs/api_keys_tab/views/success_form/types';
export interface ConnectionDetailsOpts {
    links?: ConnectionDetailsOptsLinks;
    endpoints?: ConnectionDetailsOptsEndpoints;
    apiKeys?: ConnectionDetailsOptsApiKeys;
    navigateToUrl?: (url: string) => void;
    onTelemetryEvent?: (event: ConnectionDetailsTelemetryEvents) => void;
    defaultTabId?: TabID;
}
export interface ConnectionDetailsOptsLinks {
    learnMore?: string;
}
export interface ConnectionDetailsOptsEndpoints {
    url?: string;
    id?: string;
    cloudIdLearMoreLink?: string;
}
export interface ConnectionDetailsOptsApiKeys {
    manageKeysLink?: string;
    createKey: (params: {
        name: string;
    }) => Promise<{
        apiKey: ApiKey;
    }>;
    hasPermission: () => Promise<boolean>;
}
export type ConnectionDetailsTelemetryEvent<EventId extends string, EventPayload = void> = [
    id: EventId,
    payload?: EventPayload
];
export type ConnectionDetailsTelemetryEvents = ConnectionDetailsTelemetryEvent<'learn_more_clicked'> | ConnectionDetailsTelemetryEvent<'tab_switched', {
    tab: string;
}> | ConnectionDetailsTelemetryEvent<'copy_endpoint_url_clicked'> | ConnectionDetailsTelemetryEvent<'show_cloud_id_toggled'> | ConnectionDetailsTelemetryEvent<'copy_cloud_id_clicked'> | ConnectionDetailsTelemetryEvent<'new_api_key_created'> | ConnectionDetailsTelemetryEvent<'manage_api_keys_clicked'> | ConnectionDetailsTelemetryEvent<'key_encoding_changed', {
    format: string;
}> | ConnectionDetailsTelemetryEvent<'copy_api_key_clicked', {
    format: string;
}>;
export type TabID = 'endpoints' | 'apiKeys';